UIParse/requirements.txt
PyQt6
xmltodict
pyyaml
Pygments
